refactor($compile): 메뉴, 메뉴그룹 패키지 구성 분리
