+++ src/main/java/kitchenpos/menu/menu/domain/ProfanityClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.menu.domain;
 
 public interface ProfanityClient {
     boolean containsProfanity(String text);
