+++ src/main/java/kitchenpos/menu/menu/infra/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.menu.infra;
 
+import kitchenpos.menu.menu.domain.Menu;
+import kitchenpos.menu.menu.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
