+++ src/main/java/kitchenpos/menu/menugroup/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.menugroup.domain;
 
 import java.util.List;
 import java.util.Optional;
