+++ src/test/java/kitchenpos/menu/menu/infra/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.menu.infra;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.menugroup.domain.MenuGroup;
+import kitchenpos.menu.menugroup.domain.MenuGroupRepository;
 
 import java.util.*;
