+++ src/test/java/kitchenpos/menu/menugroup/infra/InMemoryMenuRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.menugroup.infra;
 
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.menu.domain.Menu;
+import kitchenpos.menu.menu.domain.MenuRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
