+++ src/test/java/kitchenpos/orders/eatin/application/EatInOrderServiceTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.orders.eatin.application;
 
-import kitchenpos.menu.application.InMemoryMenuRepository;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.menugroup.infra.InMemoryMenuRepository;
+import kitchenpos.menu.menu.domain.MenuRepository;
 import kitchenpos.orders.eatin.domain.EatInOrder;
 import kitchenpos.orders.eatin.domain.OrderLineItem;
 import kitchenpos.orders.eatin.domain.OrderRepository;
